# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR & Payroll Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll:
    get:
      operationId: listPayroll
      tags: [payroll]
      summary: List payroll records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payroll records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
    post:
      operationId: createPayroll
      tags: [payroll]
      summary: Create payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollBody"
      responses:
        "201":
          description: Payroll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/{id}:
    put:
      operationId: updatePayroll
      tags: [payroll]
      summary: Update payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
    delete:
      operationId: deletePayroll
      tags: [payroll]
      summary: Delete payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Submit leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveBody"
      responses:
        "201":
          description: Leave request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}:
    put:
      operationId: updateLeave
      tags: [leaves]
      summary: Update leave status (approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    delete:
      operationId: deleteLeave
      tags: [leaves]
      summary: Delete leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Payroll:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        bonuses:
          type: number
        deductions:
          type: number
        netSalary:
          type: number
        status:
          type: string
          enum: [pending, processed, paid]
        processedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - month
        - year
        - basicSalary
        - bonuses
        - deductions
        - netSalary
        - status
        - createdAt
    CreatePayrollBody:
      type: object
      properties:
        employeeId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        basicSalary:
          type: number
        bonuses:
          type: number
        deductions:
          type: number
        status:
          type: string
          enum: [pending, processed, paid]
      required:
        - employeeId
        - month
        - year
        - basicSalary
        - bonuses
        - deductions
        - status
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
          nullable: true
        leaveType:
          type: string
          enum: [annual, sick, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: integer
          nullable: true
        reviewNote:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - leaveType
        - startDate
        - endDate
        - status
        - createdAt
    CreateLeaveBody:
      type: object
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
          nullable: true
      required:
        - employeeId
        - leaveType
        - startDate
        - endDate
    UpdateLeaveBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewNote:
          type: string
          nullable: true
      required:
        - status
...