# lib/api-zod/src/generated/types/createPayrollBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR & Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePayrollBodyStatus } from "./createPayrollBodyStatus";

export interface CreatePayrollBody {
  employeeId: number;
  month: number;
  year: number;
  basicSalary: number;
  bonuses: number;
  deductions: number;
  status: CreatePayrollBodyStatus;
}