# lib/api-zod/src/generated/types/updateLeaveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR & Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLeaveBodyStatus } from "./updateLeaveBodyStatus";

export interface UpdateLeaveBody {
  status: UpdateLeaveBodyStatus;
  reviewNote?: string | null;
}